Moved the router to a separate file
