+++ backend/app/main.py
@@ -1,6 +1,6 @@
 from fastapi import FastAPI
 from app.database import supabase
-from .routers import users, camera, recipes, ask, posts, AiNav
+from .routers import users, camera, recipes, ask, posts, AiNav, comments
 from fastapi.middleware.cors import CORSMiddleware
 import os
 import uvicorn
@@ -12,6 +12,7 @@ app.include_router(recipes.router)
 app.include_router(ask.router)
 app.include_router(posts.router)
 app.include_router(AiNav.router)
+app.include_router(comments.router)
 origins = [
     "http://localhost:3000"
 ]
